Turned image caching back on.

+++ controller.rb
@@ -241,17 +241,16 @@ class Controller < Sinatra::Base
     file_path_tmp = "#{file_path}tmp"
     marker_path = File.join Controller.root, "public", "img", "player-icon-" + params[:team] + ".png"
 
-    #if File.exist?(file_path)
-    #  send_file file_path
-    #else
+    if File.exist?(file_path)
+      send_file file_path
+    else
       file_path_1 = File.join Controller.root, "public", "characters", params[:i1]+".png"
       file_path_2 = File.join Controller.root, "public", "characters", params[:i2]+".png"
 
-      puts "convert \( #{marker_path} \( -geometry +11+6 -compose Over \( #{file_path_2} -resize 130% \) \) -composite \) \( -geometry +2+6 -compose Over \( #{file_path_1} -resize 130% \) \) -composite #{file_path_tmp}"
       `convert \\( #{marker_path} \\( -geometry +11+6 -compose Over \\( #{file_path_2} -resize 130% \\) \\) -composite \\) \\( -geometry +2+6 -compose Over \\( #{file_path_1} -resize 130% \\) \\) -composite #{file_path_tmp}`
       FileUtils.mv file_path_tmp, file_path
       send_file file_path
-    #end
+    end
   end
 
   post '/contact_submit' do
